Add initial implementation of HTTP handlers and main server setup
